Publish tarball with single JS file build to npm (#3866)
* Remove unneeded package.json fields from dist

We don't need dependencies and devDependencies in the package.json within the Yarn distributable, as the dependencies are bundled with Yarn itself. We also don't need "scripts" or "jest" as they're purely for development

* Add script for publishing to npm

* Add Jenkins job for publishing to npm
