+++ scripts/build-dist.sh
@@ -27,6 +27,7 @@ eval $system_yarn run build-bundle
 chmod +x artifacts/*.js
 
 cp package.json dist/
+cp README.md dist/
 cp LICENSE dist/
 cp artifacts/yarn-legacy-*.js dist/lib/yarn-cli.js
 cp bin/yarn-bundle-entry.js dist/bin/yarn.js
@@ -35,5 +36,5 @@ cp bin/{yarn,yarnpkg,*.cmd} dist/bin/
 cp node_modules/v8-compile-cache/v8-compile-cache.js dist/lib/v8-compile-cache.js
 
 version=`exec $dist_yarn --version`
-./scripts/set-installation-method.js $(readlink -f dist/package.json) tar
+./scripts/update-dist-manifest.js $(readlink -f dist/package.json) tar
 tar -cvzf artifacts/yarn-v$version.tar.gz dist/*
